+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to Semantic Versioning. Use prefixes (Added, Changed, Deprecated, Removed, Fixed, Security) for each change.
 
+## 1.1.0 - 07/16/20
+
+- `Added` MultiTypeInput to formik, and added allowable types prop
+- `Added` MultiSelectTypeInput component
+
 ## 1.0.195 - 07/16/20
 
 - `Added` activity icon for nextGen